rio: use 32b offsets/lengths
decouple records' and blocks' sizes (32b) from expected file offsets
and sizes (64b.)

A single record is meant to store at most a single HEP event.
Such events should not be greater than 4Gb (uncompressed) in size, so
there's no need to use uint64.
Go with uint32.

+++ block.go
@@ -50,7 +50,7 @@ func (blk *Block) Write(data interface{}) error {
 	}
 
 	blk.raw.Data = buf.Bytes()
-	blk.raw.Header.Len = uint64(len(blk.raw.Data))
+	blk.raw.Header.Len = uint32(len(blk.raw.Data))
 	return nil
 }
